+++ pkg/kn/commands/revision/revision_describe.go
@@ -12,58 +12,58 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package commands
+package revision
 
 import (
-	"fmt"
+	"errors"
 
+	"github.com/knative/client/pkg/kn/commands"
 	"github.com/spf13/cobra"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
 )
 
-// NewServiceGetCommand represents 'kn service get' command
-func NewServiceGetCommand(p *KnParams) *cobra.Command {
-	serviceGetFlags := NewServiceGetFlags()
-
-	serviceGetCommand := &cobra.Command{
-		Use:   "get",
-		Short: "Get available services.",
+func NewRevisionDescribeCommand(p *commands.KnParams) *cobra.Command {
+	revisionDescribePrintFlags := genericclioptions.NewPrintFlags("").WithDefaultOutput("yaml")
+	revisionDescribeCmd := &cobra.Command{
+		Use:   "describe NAME",
+		Short: "Describe revisions.",
 		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) < 1 {
+				return errors.New("requires the revision name.")
+			}
+
 			client, err := p.ServingFactory()
 			if err != nil {
 				return err
 			}
-			namespace, err := GetNamespace(cmd)
+
+			namespace, err := commands.GetNamespace(cmd)
 			if err != nil {
 				return err
 			}
-			service, err := client.Services(namespace).List(v1.ListOptions{})
+			revision, err := client.Revisions(namespace).Get(args[0], v1.GetOptions{})
 			if err != nil {
 				return err
 			}
-			if len(service.Items) == 0 {
-				fmt.Fprintf(cmd.OutOrStdout(), "No resources found.\n")
-				return nil
+
+			printer, err := revisionDescribePrintFlags.ToPrinter()
+			if err != nil {
+				return err
 			}
-			service.GetObjectKind().SetGroupVersionKind(schema.GroupVersionKind{
+			revision.GetObjectKind().SetGroupVersionKind(schema.GroupVersionKind{
 				Group:   "knative.dev",
 				Version: "v1alpha1",
-				Kind:    "Service"})
-
-			printer, err := serviceGetFlags.ToPrinter()
-			if err != nil {
-				return err
-			}
-
-			err = printer.PrintObj(service, cmd.OutOrStdout())
+				Kind:    "Revision"})
+			err = printer.PrintObj(revision, cmd.OutOrStdout())
 			if err != nil {
 				return err
 			}
 			return nil
 		},
 	}
-	AddNamespaceFlags(serviceGetCommand.Flags(), true)
-	serviceGetFlags.AddFlags(serviceGetCommand)
-	return serviceGetCommand
+	commands.AddNamespaceFlags(revisionDescribeCmd.Flags(), false)
+	revisionDescribePrintFlags.AddFlags(revisionDescribeCmd)
+	return revisionDescribeCmd
 }
