+++ pkg/kn/commands/service/service.go
@@ -12,18 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package commands
+package service
 
 import (
+	"github.com/knative/client/pkg/kn/commands"
 	"github.com/spf13/cobra"
 )
 
-func NewRevisionCommand(p *KnParams) *cobra.Command {
-	revisionCmd := &cobra.Command{
-		Use:   "revision",
-		Short: "Revision command group",
+func NewServiceCommand(p *commands.KnParams) *cobra.Command {
+	serviceCmd := &cobra.Command{
+		Use:   "service",
+		Short: "Service command group",
 	}
-	revisionCmd.AddCommand(NewRevisionGetCommand(p))
-	revisionCmd.AddCommand(NewRevisionDescribeCommand(p))
-	return revisionCmd
+	serviceCmd.AddCommand(NewServiceGetCommand(p))
+	serviceCmd.AddCommand(NewServiceDescribeCommand(p))
+	serviceCmd.AddCommand(NewServiceCreateCommand(p))
+	serviceCmd.AddCommand(NewServiceDeleteCommand(p))
+	serviceCmd.AddCommand(NewServiceUpdateCommand(p))
+	return serviceCmd
 }
