+++ pkg/kn/commands/service/service_update_test.go
@@ -12,19 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package commands
+package service
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"reflect"
 	"testing"
 
+	"github.com/knative/client/pkg/kn/commands"
 	servinglib "github.com/knative/client/pkg/serving"
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
-	serving "github.com/knative/serving/pkg/client/clientset/versioned/typed/serving/v1alpha1"
-	"github.com/knative/serving/pkg/client/clientset/versioned/typed/serving/v1alpha1/fake"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -37,13 +35,8 @@ func fakeServiceUpdate(original *v1alpha1.Service, args []string) (
 	updated *v1alpha1.Service,
 	output string,
 	err error) {
-
-	buf := new(bytes.Buffer)
-	fakeServing := &fake.FakeServingV1alpha1{&client_testing.Fake{}}
-	cmd := NewKnCommand(KnParams{
-		Output:         buf,
-		ServingFactory: func() (serving.ServingV1alpha1Interface, error) { return fakeServing, nil },
-	})
+	knParams := &commands.KnParams{}
+	cmd, fakeServing, buf := commands.CreateTestKnCommand(NewServiceCommand(knParams), knParams)
 	fakeServing.AddReactor("update", "*",
 		func(a client_testing.Action) (bool, runtime.Object, error) {
 			updateAction, ok := a.(client_testing.UpdateAction)
