+++ pkg/kn/commands/version.go
@@ -16,6 +16,7 @@ package commands
 
 import (
 	"fmt"
+
 	"github.com/spf13/cobra"
 )
