+++ pkg/kn/core/root.go
@@ -12,44 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package commands
+package core
 
 import (
 	"flag"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 
-	serving "github.com/knative/serving/pkg/client/clientset/versioned/typed/serving/v1alpha1"
+	"github.com/knative/client/pkg/kn/commands"
+	"github.com/knative/client/pkg/kn/commands/revision"
+	"github.com/knative/client/pkg/kn/commands/service"
 	homedir "github.com/mitchellh/go-homedir"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/oidc"
-	"k8s.io/client-go/tools/clientcmd"
 )
 
-var cfgFile string
-var kubeCfgFile string
-
-// Parameters for creating commands. Useful for inserting mocks for testing.
-type KnParams struct {
-	Output         io.Writer
-	ServingFactory func() (serving.ServingV1alpha1Interface, error)
-}
-
-func (c *KnParams) Initialize() {
-	if c.ServingFactory == nil {
-		c.ServingFactory = GetConfig
-	}
-}
-
 // rootCmd represents the base command when called without any subcommands
-func NewKnCommand(params ...KnParams) *cobra.Command {
-	var p *KnParams
+func NewKnCommand(params ...commands.KnParams) *cobra.Command {
+	var p *commands.KnParams
 	if len(params) == 0 {
-		p = &KnParams{}
+		p = &commands.KnParams{}
 	} else if len(params) == 1 {
 		p = &params[0]
 	} else {
@@ -78,12 +63,13 @@ Eventing: Manage event subscriptions and channels. Connect up event sources.`,
 	if p.Output != nil {
 		rootCmd.SetOutput(p.Output)
 	}
-	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.kn.yaml)")
-	rootCmd.PersistentFlags().StringVar(&kubeCfgFile, "kubeconfig", "", "kubectl config file (default is $HOME/.kube/config)")
-	rootCmd.AddCommand(NewServiceCommand(p))
-	rootCmd.AddCommand(NewRevisionCommand(p))
-	rootCmd.AddCommand(NewCompletionCommand(p))
-	rootCmd.AddCommand(NewVersionCommand(p))
+	rootCmd.PersistentFlags().StringVar(&commands.CfgFile, "config", "", "config file (default is $HOME/.kn.yaml)")
+	rootCmd.PersistentFlags().StringVar(&commands.KubeCfgFile, "kubeconfig", "", "kubectl config file (default is $HOME/.kube/config)")
+
+	rootCmd.AddCommand(service.NewServiceCommand(p))
+	rootCmd.AddCommand(revision.NewRevisionCommand(p))
+	rootCmd.AddCommand(commands.NewCompletionCommand(p))
+	rootCmd.AddCommand(commands.NewVersionCommand(p))
 
 	// For glog parse error.
 	flag.CommandLine.Parse([]string{})
@@ -97,26 +83,26 @@ func InitializeConfig() {
 }
 
 func initKubeConfig() {
-	if kubeCfgFile != "" {
+	if commands.KubeCfgFile != "" {
 		return
 	}
 	if kubeEnvConf, ok := os.LookupEnv("KUBECONFIG"); ok {
-		kubeCfgFile = kubeEnvConf
+		commands.KubeCfgFile = kubeEnvConf
 	} else {
 		home, err := homedir.Dir()
 		if err != nil {
 			fmt.Fprintln(os.Stderr, err)
 			os.Exit(1)
 		}
-		kubeCfgFile = filepath.Join(home, ".kube", "config")
+		commands.KubeCfgFile = filepath.Join(home, ".kube", "config")
 	}
 }
 
 // initConfig reads in config file and ENV variables if set.
 func initConfig() {
-	if cfgFile != "" {
+	if commands.CfgFile != "" {
 		// Use config file from the flag.
-		viper.SetConfigFile(cfgFile)
+		viper.SetConfigFile(commands.CfgFile)
 	} else {
 		// Find home directory.
 		home, err := homedir.Dir()
@@ -136,16 +122,4 @@ func initConfig() {
 	if err := viper.ReadInConfig(); err == nil {
 		fmt.Fprintln(os.Stderr, "Using config file:", viper.ConfigFileUsed())
 	}
-}
-
-func GetConfig() (serving.ServingV1alpha1Interface, error) {
-	config, err := clientcmd.BuildConfigFromFlags("", kubeCfgFile)
-	if err != nil {
-		return nil, err
-	}
-	client, err := serving.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-	return client, nil
 }
